// src/generated/api/types/movieReviews200ResultsItem.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * tmdb-api
 * OpenAPI spec version: 3
 */
import type { MovieReviews200ResultsItemAuthorDetails } from './movieReviews200ResultsItemAuthorDetails.ts';

export type MovieReviews200ResultsItem = {
  author?: string;
  author_details?: MovieReviews200ResultsItemAuthorDetails;
  content?: string;
  created_at?: string;
  id?: string;
  updated_at?: string;
  url?: string;
};